chore: rename helpers
